technical update for automatic problem type detection added

+++ template/template_types.h
@@ -1,3 +1,3 @@
-typedef float EVALUATION_TYPE;
+typedef long EVALUATION_TYPE;
 typedef int GENOME_TYPE;
-typedef float FITNESS_TYPE;
+typedef int FITNESS_TYPE;
